modified:   src/access/rtsp/RtspFilter.cpp 	modified:   src/config/localConfig/LocalManager.cpp 	deleted:    src/config/localConfig/localConfig.project~ 	modified:   src/control/joiss/ControlManager.cpp 	modified:   src/control/joiss/ControlManager.h 	modified:   src/control/joiss/JoissManager.h 	modified:   src/control/joiss/MCManager.cpp 	modified:   src/control/josp/JospParser.cpp 	modified:   src/control/stream/DeviceControl.h 	modified:   src/control/stream/StreamManager.h 	modified:   src/core/j_common.h 	modified:   src/core/x_asio.cpp 	modified:   src/core/x_asio.h 	modified:   src/core/x_lock.cpp 	modified:   src/core/x_lock.h 	modified:   src/core/x_string.h 	modified:   src/files/jo_remote/JorFileReader.cpp 	modified:   src/files/jo_remote/JorFileReader.h 	modified:   src/files/josf/JofsFileReader.cpp 	modified:   src/files/josf/JofsFileReader.h 	modified:   src/host/aipstar/AipstarAdapter.h 	modified:   src/host/aipstar/AipstarStream.h 	deleted:    src/host/aipstar/aipstarHost.project~ 	modified:   src/host/aipstar/tmControlClient.h 	modified:   src/host/aipstar/tmTransDefine.h 	modified:   src/host/aironix/AironixStream.h 	modified:   src/host/aironix/dvrdvstypedef.h 	modified:   src/host/dahua/DahuaAdapter.h 	modified:   src/host/dahua/DahuaChannel.cpp 	modified:   src/host/dahua/DahuaStream.h 	modified:   src/host/hik/HikAdapter.cpp 	modified:   src/host/hik/HikAdapter.h 	modified:   src/host/hik/HikChannel.cpp 	modified:   src/host/hik/HikIntercom.cpp 	modified:   src/host/hik/HikIntercom.h 	modified:   src/host/hik/HikStream.cpp 	modified:   src/host/hik/HikStream.h 	modified:   src/host/hik/HikType.h 	modified:   src/host/hik/HikVodStream.cpp 	modified:   src/host/hik/HikVodStream.h 	modified:   src/host/jos/JospHelper.cpp 	modified:   src/joiss.cpp 	modified:   src/misc/x_http_helper.h 	modified:   src/misc/x_inet.cpp 	modified:   src/misc/x_jo_sdk.h 	modified:   src/misc/x_mux_factory.h 	modified:   src/misc/x_parser_factory.h 	modified:   src/stream_output/ps/PsMux.cpp 	modified:   src/stream_output/ps/PsMux.h 	modified:   src/stream_output/rtp/RtpMux.cpp 	modified:   src/stream_output/rtp/RtpMux.h 	modified:   src/stream_output/ts/TsMux.cpp 	modified:   src/stream_output/ts/TsMux.h 	modified:   src/utils/G722_1_1C_Float.h 	modified:   src/utils/common/x_base64.cpp 	modified:   src/utils/common/x_crc32.cpp 	modified:   src/utils/common/x_md5.cpp 	modified:   src/utils/json.h 	modified:   src/utils/sqlite3.h 	modified:   src/utils/sqlite3/shell.c 	modified:   src/utils/sqlite3/sqlite3.c 	deleted:    src/utils/sqlite3ext.h 	modified:   src/utils/tinyxml.h 	modified:   src/utils/uuid.h 	modified:   src/utils/x_base64.h 	modified:   src/utils/x_crc32.h 	modified:   src/utils/x_md5.h 	modified:   src/x_include.h 	modified:   vs2008/joiss/JoissManager.vcproj 	modified:   vs2008/joiss/StreamManager.vcproj 	modified:   vs2008/joiss/joiss.sln 	modified:   vs2008/joiss/joiss.vcproj 	modified:   vs2008/joiss/utils.vcproj

+++ src/host/dahua/DahuaChannel.cpp
@@ -109,7 +109,7 @@ int CDahuaChannel::PtzControl(int nCmd, int nParam)
 	else
 	{
 		int nSpeed = nParam / 32 + 1;
-		int nSpeed2 = sqrt(nSpeed * nSpeed / 2);
+		int nSpeed2 = sqrt((double)(nSpeed * nSpeed / 2));
 		switch (nCmd)
 		{
 		case jo_ptz_up:
